requirements.txt
flask
flask_uploads
flask_wtf
wtforms 
keras
tensorflow
numpy
imutils
opencv-python
gunicorn
boto3
azure-storage
google-cloud-storag
